First version of CT to MCNP  Matlab code
The code dicom_orienter_gui.m runs a gui using which you can modify a CT
scan and choose the radiactive seed positions inside patient body.
funally you can save the editted CT to a Mat file which can be used to
model the CT for the code MCNP5. in order to do this, you should load
the Mat file which you had saved from previous step, and the run the
code "total_catheter_simulator_last_works.m to create the input file for
MCNP5. This directory also includes some extra files for calculation of
Gamma index or source shielding effects which are being published in PMB
journal.